requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
markdown==3.5.2
MarkupSafe==2.1.3
python-dotenv==1.0.1
Werkzeug==3.0.1
WTForms==3.1.2
